got.rs: Work around assertion failure in the dynamic linker
The dynamic linker treats requests to load itself specially: rather than opening another copy, it
adds a dummy entry to the namespace.  When the program exits, _dl_fini() asserts that it will be
able to unload at least *one* fewer dynamic libraries than the number loaded, in an apparent attempt
to account for the special case.  Unfortunately, the logic for loading the dummy entry appears to
supersede the check for whether a library is already opened, so manually loading the dynamic linker
itself can result in the creation of a *duplicate* dummy entry and cause the assertion to fail!