bind_not: Add demo program showing how to convince the runtime dynamic linker to lazily update an arbitrary pointer in lieu of the GOT entry
This trick provides a superset of the functionality provided by the LD_BIND_NOT environment
variable, as described in ld.so(8).  It can be used e.g. to cause the dynamic linker to update the
entries of a shadow GOT rather than those of the real one in a doubly-indirect trampoline setup!