lrp: Add demo of a situation where a shared library defines a PT_GNU_RELRO (segment to be made read-only at link time) that includes the GOT
We obtain such a sample library by stripping the BIND_NOW flags from the system librrpreload.so.
(We also disable its constructor to prevent it from making an out-of-range syscall that causes a
crash when not caught by ptrace().)  This allows us to showcase how lazy calls cannot be resolved
and, in the case of DT_BIND_NOW, we cannot manually modify the GOT to record the link_map pointer.