posix_safety: Add snow, a demo program that works with both fully- and partially-whitelisted libc
The snow program draws a bunch of snow with a progress spinner that eventually freezes in the cold
due to a deadlock.  Unlike the as program, it doesn't have a pathological reliance on the buffer's
being shared across signal handlers, so it works without LIBGOTCHA_SHAREDLIBC being defined.

It's only when LIBGOTCHA_SHAREDLIBC is defined that libas-safe actually has to defer any signals.