Makefile: Support building against a custom glibc sysroot when invoked as $ make LDFLAGS=-L
The path to the .../lib directory must be provided along with the switch.  This feature is to
support running with an older version of glibc than is installed on the system.  As such, it
requires relinking libstd to avoid symbol version clashes.  We ask the compiler to generate an
"empty" staticlib to provide the __rust_*alloc symbols; it also includes the symbols we need from
/usr/lib/rustlib/x86_64-unknown-linux-gnu/lib/lib*.rlib, so we do not need to bother linking against
the rest of those archives.